refactor: 💡 Split into smaller modules
