+++ libs/extract/src/lib/dropdown/index.ts
@@ -1,0 +1,13 @@
+export {
+  close,
+  create,
+  open,
+  select,
+  toggle,
+} from './dropdown'
+
+export {
+  AbstractDropdown,
+  DropdownOption,
+  ExtendedDropdownOption,
+} from './types'
